refactor: create EmissionsSummary for readability
