+++ src/view/components/results/Results.tsx
@@ -1,14 +1,11 @@
 import React from "react";
 import {
-    formatBytes,
-    formatEmissions,
-} from "../../../utils/helpers/formatNumbersToString";
-import {
     HistoryContext,
     HistoryContextType,
 } from "../../provider/history/HistoryProvider";
 import { useNullSafeContext } from "../../provider/useNullSafeContext";
-import { EmissionsComparison } from "../emissions-comparison/EmissionsComparison";
+import { EmissionsComparison } from "./comparison/EmissionsComparison";
+import { EmissionsSummary } from "./summary/EmissionsSummary";
 export const Results = () => {
     const { calculationHistory } =
         useNullSafeContext<HistoryContextType>(HistoryContext);
@@ -16,18 +13,7 @@ export const Results = () => {
     return (
         calculationHistory && (
             <div className=" h-32 grid grid-cols-2 text-base bg-nyanza rounded-2xl shadow font-medium">
-                <p className="text-center flex flex-wrap content-center justify-center">
-                    {formatBytes(calculationHistory[0].bytes)}
-                    <br />
-                    {`${formatEmissions(
-                        calculationHistory[0].specificEmissions
-                    )}
-                        gCO2/GB`}
-                    <br />
-                    {`${formatEmissions(
-                        calculationHistory[0].emissions
-                    )} g of CO2`}
-                </p>
+                <EmissionsSummary />
                 <EmissionsComparison calculation={calculationHistory[0]} />
             </div>
         )
